Reimplemented TSL500Series
