+++ docs/api/instruments/santec/index.rst
@@ -9,5 +9,6 @@ This section contains specific documentation on the Santec instruments that are 
 .. toctree::
    :maxdepth: 2
 
+   tsl500series
    tsl570
    tsl550
