+++ pymeasure/instruments/santec/__init__.py
@@ -22,5 +22,6 @@
 # THE SOFTWARE.
 #
 
+from .tsl500series import TSL500Series
 from .tsl570 import TSL570
 from .tsl550 import TSL550
